Moved testclient to the dedicated folder.
