--- fleetspeak/src/client/daemonservice/testclient/testclient_launcher.py
@@ -1,32 +1,0 @@
-#!/usr/bin/env python
-# Copyright 2018 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Script that launches testclient.Loopback in another process."""
-
-import multiprocessing
-
-from absl import app
-
-from fleetspeak.src.client.daemonservice.testclient import testclient
-
-def main(argv=None):
-  del argv
-  p = multiprocessing.Process(target=testclient.Loopback)
-  p.start()
-  p.join()
-
-
-if __name__ == "__main__":
-  app.run(main)
